requirements.txt
langchain_cohere==0.1.8
starlette==0.37.2
streamlit==1.36.0
python-dotenv==1.0.1
python-socketio==5.11.3
langsmith==0.1.82
PyJWT==2.9.0
pymongo==4.8.0
pydantic==2.7.4
langchain==0.2.13
langchain-openai==0.1.13
langchain-pinecone==0.1.1
openai==1.35.7
pinecone==4.0.0
pinecone-client==3.2.2
fastapi==0.112.0
Flask==3.0.3
Flask-SocketIO==5.3.6
uvicorn==0.30.5
websocket-client==1.8.0
websockets==12.0
